Extract trimAndSumUniqMetWords
